+++ src/Config.php
@@ -22,5 +22,26 @@ class Config extends ContainerConfig
         ];
 
         $di->params['Cadre\CliAdr\Resolver']['injectionFactory'] = $di->getInjectionFactory();
+
+        $di->params['Cadre\CliAdr\Command\Help'] = [
+            'router' => $di->lazyGet('cadre:cliadr/router'),
+            'resolver' => $di->lazyGet('cadre:cliadr/resolver'),
+        ];
+
+        $di->params['Cadre\CliAdr\Command\Index'] = [
+            'router' => $di->lazyGet('cadre:cliadr/router'),
+            'resolver' => $di->lazyGet('cadre:cliadr/resolver'),
+        ];
+    }
+
+    public function modify(Container $di)
+    {
+        $adr = $di->get('cadre:cliadr/adr');
+
+        $adr->route('help', 'Cadre\CliAdr\Command\Help')
+            ->input('Cadre\CliAdr\Input\Help');
+
+        $adr->route('index', 'Cadre\CliAdr\Command\Index')
+            ->input('Cadre\CliAdr\Input\Index');
     }
 }