+++ src/Input/HelpAware.php
@@ -1,0 +1,9 @@
+<?php
+namespace Cadre\CliAdr\Input;
+
+use Aura\Cli\Help;
+
+interface HelpAware
+{
+    public function help(Help $help);
+}