+++ src/Input/Input.php
@@ -1,14 +1,15 @@
 <?php
-namespace Cadre\CliAdr;
+namespace Cadre\CliAdr\Input;
 
 use Aura\Cli\Context;
+use Aura\Cli\Help;
 use Aura\Cli\Stdio;
 
 class Input
 {
     public function __invoke(Context $context, Stdio $stdio)
     {
-        $getopt = $context->getopt(['h,help', 'v,verbose']);
+        $getopt = $context->getopt(['v,vv,vvv,verbose']);
         return [$getopt->get()];
     }
 }