+++ CHANGELOG.md
@@ -7,11 +7,20 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [0.1.0]
 ### Changed
 - Updating dev dependencies
-- 100% code coverage 
+- 100% code coverage
+
 ## [0.1.1]
 ### Added
 - `CHANGELOG.md` file
+
 ## [0.1.2]
 ### Added
 - `composer.lock` to `.gitignore`
 - Requirement for PHP >=7.0
+
+## [0.2.0]
+### Added
+- `Cadre\CliAdr\Boot` class to bootstrap system
+- Command `list` to list available commands
+- Command `help` to display help for a command
+- `Cadre\CliAdr\HelpAware` interface for inputs to configure a `Aura\Cli\Help` object.